% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/some_functions.R
\name{at_bats}
\alias{at_bats}
\title{At-bats}
\usage{
at_bats(outcomes)
}
\arguments{
\item{outcomes}{Character vector of valid outcomes for a plate appearance}
}
\description{
Given the outcomes of plate appearances, calculate player's
             at-bats.
}
\references{
\href{https://en.wikipedia.org/wiki/Baseball_statistics}{baseball statistics}
}
\seealso{
Other Batting Statistics: \code{\link{batting_avg}},
  \code{\link{on_base_pct}}, \code{\link{ops}},
  \code{\link{slugging_avg}}
}
