% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg.R
\name{avg}
\alias{avg}
\title{Batting Average}
\usage{
avg(h, ab)
}
\arguments{
\item{h}{Number of Hits the batter hit}

\item{ab}{At Bats for the batter}
}
\value{
The batter's Batting Average (AVG) is returned.
}
\description{
Batting Average
}
\examples{
avg(197, 636)
avg(h = 138, ab = 537)
}
