% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batting_statistics.R
\name{batting_avg}
\alias{batting_avg}
\title{Batting Average}
\usage{
batting_avg(outcomes)
}
\arguments{
\item{outcomes}{Character vector of valid outcomes for a plate appearance}
}
\description{
Given the outcomes of plate appearances, calculate player's
             batting average.
}
\references{
\href{https://en.wikipedia.org/wiki/Baseball_statistics}{baseball statistics}
}
\seealso{
Other Batting Statistics: 
\code{\link{at_bats}()},
\code{\link{on_base_pct}()},
\code{\link{ops}},
\code{\link{slugging_avg}()}
}
\concept{Batting Statistics}
