% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/era.R
\name{era}
\alias{era}
\title{Earned Run Average}
\usage{
era(er, ip)
}
\arguments{
\item{er}{Earned Runs surrendered by a pitcher}

\item{ip}{Innings Pitched by a Pitcher}
}
\value{
The pitcher's Earned Run Average (ERA) is returned. ERA measures how many runs are given up on a 9-inning basis.
}
\description{
Earned Run Average
}
\examples{
era(58, 228.2)
era(er = 85, ip = 197.1)
}
