% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixip.R
\name{fixip}
\alias{fixip}
\title{Fix Innings Pitched Value}
\usage{
fixip(ip)
}
\arguments{
\item{ip}{Innings Pitched by a Pitcher}
}
\value{
The Innings Pitched by a pitcher, as expressed as a fraction. Innings in baseball are expressed in terms of thirds, but in recording them, the thirds are represented by 0.1, 0.2, and 1. This function changes that in order to perform proper analysis on the IP values.
}
\description{
Fix Innings Pitched Value
}
\examples{
fixip(42.1)
fixip(118.2)
fixip(190)
}
