% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpts.R
\name{fpts}
\alias{fpts}
\title{Calculate Fantasy Points}
\usage{
fpts(
  r = 0,
  tb = 0,
  rbi = 0,
  bb = 0,
  k = 0,
  sb = 0,
  ip = NULL,
  ha = 0,
  er = 0,
  bba = 0,
  so = 0,
  w = 0,
  l = 0,
  sv = 0,
  hd = 0
)
}
\arguments{
\item{r}{Number of Runs a batter scored}

\item{tb}{Total Bases acquired by a batter}

\item{rbi}{Number of Runs Batted In a batter hit}

\item{bb}{Number of times a batter Walked}

\item{k}{Number of times a batter struck out}

\item{sb}{Stolen Bases acquired by a batter}

\item{ip}{Innings Pitched by a Pitcher}

\item{ha}{Hits Allowed by a Pitcher}

\item{er}{Earned Runs allowed by a Pitcher}

\item{bba}{Walks Allowed by a Pitcher}

\item{so}{Number of Strikeouts thrown by a Pitcher}

\item{w}{Number of Wins accumulated by a Pitcher}

\item{l}{Number of Losses accumulated by a Pitcher}

\item{sv}{Number of Saves accumulated by a Pitcher}

\item{hd}{Number of Holds accumulated by a Pitcher}
}
\value{
The total Fantasy Points scored by a player is returned. Weights of values are based on the default fantasy scoring settings in an ESPN Fantasy Baseball league in the year 2025.
}
\description{
Calculate Fantasy Points
}
\examples{
fpts(ip = 6.1, ha = 7, er = 1, bba = 1, so = 9, w = 1)
fpts(r = 128, tb = 328, rbi = 109, bb = 129, sb = 7, k = 119)
}
