% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obp.R
\name{obp}
\alias{obp}
\title{Calculate a Batter's On Base Percentage}
\usage{
obp(h, bb, hbp, sf, ab)
}
\arguments{
\item{h}{Number of Hits the batter hit}

\item{bb}{Number of times the batter Walked}

\item{hbp}{Number of Times the batter was Hit by a Pitch}

\item{sf}{Number of Sacrifice Flies hit by the batter}

\item{ab}{At Bats for the batter}
}
\value{
The batter's On Base Percentage (OBP) is printed. OBP measures how often a batter reaches base.
}
\description{
Calculate a Batter's On Base Percentage
}
\examples{
obp(176, 75, 18, 3, 603)
obp(h = 197, bb = 81, hbp = 6, sf = 5, ab = 636)
}
