% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/some_functions.R
\name{on_base_pct}
\alias{on_base_pct}
\title{On Base Percentage}
\usage{
on_base_pct(outcomes)
}
\arguments{
\item{outcomes}{Character vector of valid outcomes for a plate appearance}
}
\description{
Given the outcomes of plat appearances, calculate player's
             on base percentage.
}
\references{
\href{https://en.wikipedia.org/wiki/Baseball_statistics}{baseball statistics}
}
\seealso{
Other Batting Statistics: \code{\link{at_bats}},
  \code{\link{batting_avg}}, \code{\link{ops}},
  \code{\link{slugging_avg}}
}
