% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{ops}
\alias{ops}
\title{On Base Plus Slugging Percentage}
\usage{
ops(h, bb, hbp, sf, tb, ab)
}
\arguments{
\item{h}{Number of Hits the batter hit}

\item{bb}{Number of times the batter Walked}

\item{hbp}{Number of Times the batter was Hit by a Pitch}

\item{sf}{Number of Sacrifice Flies hit by the batter}

\item{tb}{Total Bases accumulated by the batter}

\item{ab}{At Bats for the batter}
}
\value{
The batter's On Base plus Slugging Percentage (OPS) is returned. OPS is a combination of OBP and SLG and can also be calculated by summing them together. An OPS of .700 is usually an average value, while an OPS of .900 or above is considered to be extraordinary.
}
\description{
On Base Plus Slugging Percentage
}
\examples{
ops(176, 75, 18, 5, 334, 603)
ops(h = 197, bb = 81, hbp = 6, sf = 5, tb = 411, ab = 636)
}
