% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slg.R
\name{slg}
\alias{slg}
\title{Slugging Percentage}
\usage{
slg(tb, ab)
}
\arguments{
\item{tb}{Total Bases accumulated by the batter}

\item{ab}{At Bats for the batter}
}
\value{
The batter's Slugging percentage (SLG)  is returned. Slugging percentage measures the efficiency of the batter's hits, with extra weight given to more valuable hits. Slugging percentage can also be calculated by replacing Total Bases (TB) with Singles + 2 x Doubles + 3 x Triples + 4 x Home Runs.
}
\description{
Slugging Percentage
}
\examples{
slg(334, 603)
slg(tb = 411, ab = 636)
}
