% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/some_functions.R
\name{slugging_avg}
\alias{slugging_avg}
\title{Slugging average}
\usage{
slugging_avg(outcomes)
}
\arguments{
\item{outcomes}{Character vector of valid outcomes for a plate appearance}
}
\description{
Given the outcomes of plat appearances, calculate player's
             slugging average.
}
\references{
\href{https://en.wikipedia.org/wiki/Baseball_statistics}{baseball statistics}
}
\seealso{
Other Batting Statistics: \code{\link{at_bats}},
  \code{\link{batting_avg}}, \code{\link{on_base_pct}},
  \code{\link{ops}}
}
