% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statChart.R
\name{statChart}
\alias{statChart}
\title{Statistics Chart}
\usage{
statChart(df, command, n)
}
\arguments{
\item{df}{The data frame used for the creation of the chart}

\item{command}{The statistic desired to be visualized}

\item{n}{The number of players desired to be visualized}
}
\value{
Creates a bar graph showcasing the top N players at a chosen statistic from the baseballStats package. Statistics available are avg, era, fpts, obp, ops, slg, and whip.
}
\description{
Statistics Chart
}
\examples{
statChart(hitters, "slg", n = 10)
statChart(pitchers, "era", n = 15)
}
