% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whip.R
\name{whip}
\alias{whip}
\title{Walks Plus Hits Per Innings Pitched}
\usage{
whip(bba, ha, ip)
}
\arguments{
\item{bba}{Walks Allowed by a Pitcher}

\item{ha}{Hits Allowed by a Pitcher}

\item{ip}{Innings Pitched by a Pitcher}
}
\value{
The pitcher's Walks plus Hits per Innings Pitched (WHIP) is returned. WHIP measures a pitcher's tendency to allow batters to reach base.
}
\description{
Walks Plus Hits Per Innings Pitched
}
\examples{
whip(64, 112, 180.2)
whip(bba = 99, ha = 115, ip = 180)
}
