% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batting_statistics.R
\name{ops}
\alias{ops}
\alias{OPS}
\title{On-base Percentage}
\usage{
OPS(outcomes)
}
\arguments{
\item{outcomes}{Character vector of valid outcomes for a plate appearance}
}
\description{
Given the outcomes of plat appearances, calculate player's
             on base plus slugging (OPS).
}
\references{
\href{https://en.wikipedia.org/wiki/Baseball_statistics}{baseball statistics}
}
\seealso{
Other Batting Statistics: 
\code{\link{at_bats}()},
\code{\link{batting_avg}()},
\code{\link{on_base_pct}()},
\code{\link{slugging_avg}()}
}
\concept{Batting Statistics}
